test: move some browser window specs to the main process (#19089)
